refactor(docs): moved docs outside of next.js project
