+++ docs/pages/home.stories.mdx
@@ -45,6 +45,6 @@ The following component was created following the guidelines defined by code cha
   - **Pattern**: Composition Pattern
   - **Additional Information**:
     - Used Context API and custom hooks to share and read information between components to avoid prop drilling.
-    - Extended HTML Element attributes of elements used within parent tags to allow components to receive more attributes than set by the component. This allows working with Microdata when creating FAQs.
+    - Extended Accordion components to allow passing their respective HTML Attributes to work with Microdata.
     - Added `ARIA` attributes for accessibility
